+++ packages/assets-controllers/src/NftDetectionController.ts
@@ -33,6 +33,7 @@ import {
   type NftControllerState,
   type NftMetadata,
 } from './NftController';
+import type { NetworkControllerFindNetworkClientIdByChainIdAction } from '../../network-controller/src/NetworkController';
 
 const controllerName = 'NftDetectionController';
 
@@ -43,7 +44,8 @@ export type AllowedActions =
   | NetworkControllerGetStateAction
   | NetworkControllerGetNetworkClientByIdAction
   | PreferencesControllerGetStateAction
-  | AccountsControllerGetSelectedAccountAction;
+  | AccountsControllerGetSelectedAccountAction
+  | NetworkControllerFindNetworkClientIdByChainIdAction;
 
 export type AllowedEvents =
   | PreferencesControllerStateChangeEvent
@@ -794,11 +796,14 @@ export class NftDetectionController extends BaseController<
               collection && { collection },
               chainId && { chainId },
             );
-            await this.#addNft(contract, tokenId, {
+            const networkClientId = this.messagingSystem.call(
+              'NetworkController:findNetworkClientIdByChainId',
+              toHex(chainId),
+            );
+            await this.#addNft(contract, tokenId, networkClientId, {
               nftMetadata,
               userAddress,
               source: Source.Detected,
-              chainId: toHex(chainId),
             });
           }
         });