+++ packages/bridge-controller/src/bridge-controller.ts
@@ -3,11 +3,10 @@ import { Contract } from '@ethersproject/contracts';
 import { Web3Provider } from '@ethersproject/providers';
 import type { StateMetadata } from '@metamask/base-controller';
 import type { ChainId, TraceCallback } from '@metamask/controller-utils';
-import { SolScope } from '@metamask/keyring-api';
 import { abiERC20 } from '@metamask/metamask-eth-abis';
 import type { NetworkClientId } from '@metamask/network-controller';
 import { StaticIntervalPollingController } from '@metamask/polling-controller';
-import type { TransactionParams } from '@metamask/transaction-controller';
+import type { TransactionController } from '@metamask/transaction-controller';
 import type { CaipAssetType } from '@metamask/utils';
 import { numberToHex, type Hex } from '@metamask/utils';
 
@@ -68,6 +67,10 @@ import type {
 } from './utils/metrics/types';
 import { type CrossChainSwapsEventProperties } from './utils/metrics/types';
 import { isValidQuoteRequest } from './utils/quote';
+import {
+  getFeeForTransactionRequest,
+  getMinimumBalanceForRentExemptionRequest,
+} from './utils/snaps';
 
 const metadata: StateMetadata<BridgeControllerState> = {
   quoteRequest: {
@@ -99,6 +102,10 @@ const metadata: StateMetadata<BridgeControllerState> = {
     anonymous: false,
   },
   assetExchangeRates: {
+    persist: false,
+    anonymous: false,
+  },
+  minimumBalanceForRentExemptionInLamports: {
     persist: false,
     anonymous: false,
   },
@@ -138,10 +145,7 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
 
   readonly #clientId: string;
 
-  readonly #getLayer1GasFee: (params: {
-    transactionParams: TransactionParams;
-    chainId: ChainId;
-  }) => Promise<string>;
+  readonly #getLayer1GasFee: typeof TransactionController.prototype.getLayer1GasFee;
 
   readonly #fetchFn: FetchFunction;
 
@@ -172,10 +176,7 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
     messenger: BridgeControllerMessenger;
     state?: Partial<BridgeControllerState>;
     clientId: BridgeClientId;
-    getLayer1GasFee: (params: {
-      transactionParams: TransactionParams;
-      chainId: ChainId;
-    }) => Promise<string>;
+    getLayer1GasFee: typeof TransactionController.prototype.getLayer1GasFee;
     fetchFn: FetchFunction;
     config?: {
       customBridgeApiBaseUrl?: string;
@@ -249,6 +250,15 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
       ...DEFAULT_BRIDGE_CONTROLLER_STATE.quoteRequest,
       ...paramsToUpdate,
     };
+
+    if (
+      paramsToUpdate.srcChainId &&
+      paramsToUpdate.srcChainId !== this.state.quoteRequest.srcChainId
+    ) {
+      await this.#setMinimumBalanceForRentExemptionInLamports(
+        paramsToUpdate.srcChainId,
+      );
+    }
 
     this.update((state) => {
       state.quoteRequest = updatedQuoteRequest;
@@ -423,6 +433,8 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
         DEFAULT_BRIDGE_CONTROLLER_STATE.quotesRefreshCount;
       state.assetExchangeRates =
         DEFAULT_BRIDGE_CONTROLLER_STATE.assetExchangeRates;
+      state.minimumBalanceForRentExemptionInLamports =
+        DEFAULT_BRIDGE_CONTROLLER_STATE.minimumBalanceForRentExemptionInLamports;
     });
   };
 
@@ -497,6 +509,9 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
         },
         fetchQuotes,
       );
+      await this.#setMinimumBalanceForRentExemptionInLamports(
+        updatedQuoteRequest.srcChainId,
+      );
     } catch (error) {
       const isAbortError = (error as Error).name === 'AbortError';
       const isAbortedDueToReset = error === RESET_STATE_ABORT_MESSAGE;
@@ -553,38 +568,83 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
     });
 
     // Only append L1 gas fees if all quotes are for either optimism or base
-    if (!hasInvalidQuotes) {
-      return await Promise.all(
-        quotes.map(async (quoteResponse) => {
-          const { quote, trade, approval } = quoteResponse;
-          const chainId = numberToHex(quote.srcChainId) as ChainId;
-
-          const getTxParams = (txData: TxData) => ({
-            from: txData.from,
-            to: txData.to,
-            value: txData.value,
-            data: txData.data,
-            gasLimit: txData.gasLimit?.toString(),
-          });
-          const approvalL1GasFees = approval
-            ? await this.#getLayer1GasFee({
-                transactionParams: getTxParams(approval),
-                chainId,
-              })
-            : '0';
-          const tradeL1GasFees = await this.#getLayer1GasFee({
-            transactionParams: getTxParams(trade),
-            chainId,
-          });
-          return {
-            ...quoteResponse,
-            l1GasFeesInHexWei: sumHexes(approvalL1GasFees, tradeL1GasFees),
-          };
-        }),
-      );
+    if (hasInvalidQuotes) {
+      return undefined;
     }
 
-    return undefined;
+    const l1GasFeePromises = Promise.allSettled(
+      quotes.map(async (quoteResponse) => {
+        const { quote, trade, approval } = quoteResponse;
+        const chainId = numberToHex(quote.srcChainId) as ChainId;
+
+        const getTxParams = (txData: TxData) => ({
+          from: txData.from,
+          to: txData.to,
+          value: txData.value,
+          data: txData.data,
+          gasLimit: txData.gasLimit?.toString(),
+        });
+        const approvalL1GasFees = approval
+          ? await this.#getLayer1GasFee({
+              transactionParams: getTxParams(approval),
+              chainId,
+            })
+          : '0x0';
+        const tradeL1GasFees = await this.#getLayer1GasFee({
+          transactionParams: getTxParams(trade),
+          chainId,
+        });
+
+        if (approvalL1GasFees === undefined || tradeL1GasFees === undefined) {
+          return undefined;
+        }
+
+        return {
+          ...quoteResponse,
+          l1GasFeesInHexWei: sumHexes(approvalL1GasFees, tradeL1GasFees),
+        };
+      }),
+    );
+
+    const quotesWithL1GasFees = (await l1GasFeePromises).reduce<
+      (QuoteResponse & L1GasFees)[]
+    >((acc, result) => {
+      if (result.status === 'fulfilled' && result.value) {
+        acc.push(result.value);
+      } else if (result.status === 'rejected') {
+        console.error('Error calculating L1 gas fees for quote', result.reason);
+      }
+      return acc;
+    }, []);
+
+    return quotesWithL1GasFees;
+  };
+
+  readonly #setMinimumBalanceForRentExemptionInLamports = async (
+    srcChainId: GenericQuoteRequest['srcChainId'],
+  ) => {
+    const selectedAccount = this.#getMultichainSelectedAccount();
+
+    try {
+      if (isSolanaChainId(srcChainId) && selectedAccount?.metadata?.snap?.id) {
+        const fees = (await this.messagingSystem.call(
+          'SnapController:handleRequest',
+          getMinimumBalanceForRentExemptionRequest(
+            selectedAccount.metadata.snap?.id,
+          ),
+        )) as string;
+        this.update((state) => {
+          state.minimumBalanceForRentExemptionInLamports = fees;
+        });
+        return;
+      }
+    } catch (error) {
+      console.error('Error setting minimum balance for rent exemption', error);
+    }
+    this.update((state) => {
+      state.minimumBalanceForRentExemptionInLamports =
+        DEFAULT_BRIDGE_CONTROLLER_STATE.minimumBalanceForRentExemptionInLamports;
+    });
   };
 
   readonly #appendSolanaFees = async (
@@ -597,7 +657,7 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
       return undefined;
     }
 
-    return await Promise.all(
+    const solanaFeePromises = Promise.allSettled(
       quotes.map(async (quoteResponse) => {
         const { trade } = quoteResponse;
         const selectedAccount = this.#getMultichainSelectedAccount();
@@ -605,18 +665,10 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
         if (selectedAccount?.metadata?.snap?.id && typeof trade === 'string') {
           const { value: fees } = (await this.messagingSystem.call(
             'SnapController:handleRequest',
-            {
-              snapId: selectedAccount.metadata.snap?.id as never,
-              origin: 'metamask',
-              handler: 'onRpcRequest' as never,
-              request: {
-                method: 'getFeeForTransaction',
-                params: {
-                  transaction: trade,
-                  scope: SolScope.Mainnet,
-                },
-              },
-            },
+            getFeeForTransactionRequest(
+              selectedAccount.metadata.snap?.id,
+              trade,
+            ),
           )) as { value: string };
 
           return {
@@ -627,6 +679,19 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
         return quoteResponse;
       }),
     );
+
+    const quotesWithSolanaFees = (await solanaFeePromises).reduce<
+      (QuoteResponse & SolanaFees)[]
+    >((acc, result) => {
+      if (result.status === 'fulfilled' && result.value) {
+        acc.push(result.value);
+      } else if (result.status === 'rejected') {
+        console.error('Error calculating solana fees for quote', result.reason);
+      }
+      return acc;
+    }, []);
+
+    return quotesWithSolanaFees;
   };
 
   #getMultichainSelectedAccount() {
@@ -717,6 +782,12 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
           ...baseProperties,
         };
       case UnifiedSwapBridgeEventName.QuotesRequested:
+        return {
+          ...this.#getRequestParams(),
+          ...this.#getRequestMetadata(),
+          has_sufficient_funds: !this.state.quoteRequest.insufficientBal,
+          ...baseProperties,
+        };
       case UnifiedSwapBridgeEventName.QuoteError:
         return {
           ...this.#getRequestParams(),