+++ packages/profile-sync-controller/src/sdk/authentication.test.ts
@@ -244,8 +244,6 @@ describe('Authentication - SRP Flow - getAccessToken() & getUserProfile()', () =
         mockOAuth2TokenUrl: {
           status: 400,
           body: {
-            // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-
             error_description: 'invalid JWT token',
             error: 'invalid_request',
           },
@@ -421,8 +419,6 @@ describe('Authentication - SIWE Flow - getAccessToken(), getUserProfile(), signM
         mockOAuth2TokenUrl: {
           status: 400,
           body: {
-            // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-
             error_description: 'invalid JWT token',
             error: 'invalid_request',
           },